moved and renamed classes to Tests.Helper Project
