--- Tests/Tests.Helper/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace DsuDev.BusinessDays.Tests.Helper
-{
-    public class Class1
-    {
-    }
-}
