+++ Tests/Tests.Helper/TestsDataMembers/DataProviderTestData.cs
@@ -1,20 +1,21 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using AutoMapper;
-using DsuDev.BusinessDays.Services.Interfaces.FileHandling;
+using DsuDev.BusinessDays.DataAccess;
 using Moq;
+using DbModels = DsuDev.BusinessDays.DataAccess.Models;
 
-namespace DsuDev.BusinessDays.Services.Tests.TestsDataMembers
+namespace DsuDev.BusinessDays.Tests.Helper.TestsDataMembers
 {
-    public class FileLoaderTestData : IEnumerable<object[]>
+    public class DataProviderTestData : IEnumerable<object[]>
     {
         /// <inheritdoc />
         public IEnumerator<object[]> GetEnumerator()
         {
             var mockMapper = new Mock<IMapper>();
-            var mockFileReadingManager = new Mock<IFileReadingManager>();
+            var mockHolidayRepository = new Mock<IRepository<DbModels.Holiday>>();
 
-            yield return new object[] { null, mockFileReadingManager.Object};
+            yield return new object[] { null, mockHolidayRepository.Object };
             yield return new object[] { mockMapper.Object, null };
         }
