+++ Tests/UT/Services.Tests/DirectoryHelperTests.cs
@@ -1,6 +1,6 @@
-﻿using FluentAssertions;
+﻿using DsuDev.BusinessDays.Common.Tools.FluentBuilders;
+using FluentAssertions;
 using System;
-using DsuDev.BusinessDays.Common.Tools.FluentBuilders;
 using Xunit;
 
 namespace DsuDev.BusinessDays.Services.Tests
