+++ Tests/UT/Services.Tests/DsuDev.BusinessDays.Services.Tests.csproj
@@ -21,6 +21,7 @@
     <ProjectReference Include="..\..\..\Source\Common.Tools\DsuDev.BusinessDays.Common.Tools.csproj" />
     <ProjectReference Include="..\..\..\Source\Domain\DsuDev.BusinessDays.Domain.csproj" />
     <ProjectReference Include="..\..\..\Source\Services\DsuDev.BusinessDays.Services.csproj" />
+    <ProjectReference Include="..\..\Tests.Helper\DsuDev.BusinessDays.Tests.Helper.csproj" />
   </ItemGroup>
 
 </Project>
