+++ Tests/UT/Services.Tests/FileHandling/FileReadingManagerTests.cs
@@ -1,11 +1,11 @@
-﻿using System;
-using DsuDev.BusinessDays.Common.Constants;
+﻿using DsuDev.BusinessDays.Common.Constants;
 using DsuDev.BusinessDays.Common.Tools.SampleGenerators;
 using DsuDev.BusinessDays.Services.FileHandling;
 using DsuDev.BusinessDays.Services.Interfaces.FileHandling;
-using DsuDev.BusinessDays.Services.Tests.TestsDataMembers;
+using DsuDev.BusinessDays.Tests.Helper.TestsDataMembers.FileHandling;
 using FluentAssertions;
 using Moq;
+using System;
 using Xunit;
 
 namespace DsuDev.BusinessDays.Services.Tests.FileHandling
